python django/sms_api/requirements.txt
beautifulsoup4==4.7.1
certifi==2018.11.29
chardet==3.0.4
dj-database-url==0.5.0
Django==2.1.7
django-heroku==0.3.1
gunicorn==19.9.0
idna==2.8
psycopg2==2.7.7
pycryptodome==3.7.3
pytz==2018.9
requests==2.21.0
soupsieve==1.8
urllib3==1.24.1
whitenoise==4.1.2
